1) Match history is migrated to Paging 3 library 2) Champions, items, and summoner spells moved to Data Dragon representation (also retrofit service, also room support) 3) Re-worked match history viewmodel & the adapter 4) Excracted card view holder without disposer 5) Match history, match, team, and participant models are migrated to direct RIOT API (Retrofit/GSON) 6) Removed unused factory methods from the Repository 7) Removed unused match result presentation 8) Added livedata to flowable tool converter 9) Added caching to all RIOT icon singles 10) Adjusted match card layout and extracted the icon dimensions 11) Component version upgrade in gradle build

+++ app/src/main/java/com/tsarsprocket/reportmid/riotapi/matchV4/ParticipantDto.kt
@@ -9,8 +9,8 @@ data class ParticipantDto(
     @SerializedName("stats") val stats: ParticipantStatsDto,
     @SerializedName("teamId") val teamId: Int, // 100 - blue, 200 - red
     @SerializedName("timeline") val timeline: ParticipantTimelineDto,
-    @SerializedName("spell1Id") val spell1Id: Int,
-    @SerializedName("spell2Id") val spell2Id: Int,
+    @SerializedName("spell1Id") val spell1Id: Long,
+    @SerializedName("spell2Id") val spell2Id: Long,
     @SerializedName("highestAchievedSeasonTier") val highestAchievedSeasonTier: String,
     @SerializedName("masteries") val masteries: List<MasteryDto>
 )
